Allow for using "tap dance" on keyboard
This is used to allow me to assign keys to assign a function to the
case when I press the same key quickly, twice in a row.
For example, when the dash/minus key is tapped once, it sends the
minus character. When tapped twice quickly, it sends an equals-sign.
In addition, this works with the shift modifier, so if I double-tap
the dash key while shift is held, it will produce an addition symbol.